\docType{package}
\name{dmetatools-package}
\alias{dmetatools-package}
\alias{dmetatools}
\title{The 'dmetatools' package.}
\description{
A R package for computational tools of meta-analysis of diagnostic accuracy test. This package enables computations of confidence interval for AUC of summary ROC curve and some related AUC-based inference methods.
}
\references{
Noma, H. and Matsushima, Y. (2020). 
Bootstrap confidence interval for AUC of the summary ROC curve and some related methods for meta-analysis of diagnostic accuracy studies. 
Under Preparation.
}
