\docType{package}
\name{dmetatools-package}
\alias{dmetatools-package}
\title{The 'dmetatools' package.}
\description{
Computational tools for meta-analysis of diagnostic accuracy test. Bootstrap-based computational methods of the confidence interval for AUC of summary ROC curve and some related AUC-based inference methods are available.
}
\author{Hisashi Noma <noma@ism.ac.jp>}
\references{
Noma, H., Matsushima, Y., and Ishii, R. (2021). 
Confidence interval for the AUC of SROC curve and some related methods using bootstrap for meta-analysis of diagnostic accuracy studies. 
\emph{Communications in Statistics: Case Studies and Data Analysis} \strong{7}: 344-358.
\doi{10.1080/23737484.2021.1894408}
}
