\docType{data}
\name{asthma}
\alias{asthma}
\title{Korevaar et al. (2015)'s data of minimally invasive markers for detection of airway eosinophilia in asthma}
\format{A data frame with 12 rows and 4 variables}
\usage{
data(asthma)
}
\description{
\itemize{
  \item \code{TP}: A vector of the number of true positives (TP)
  \item \code{FP}: A vector of the number of false positives (FP)
  \item \code{FN}: A vector of the number of false negatives (FN)
  \item \code{TN}: A vector of the number of true negatives (TN)
}
}
\references{
Korevaar, D. A., Westerhof, G. A., Wang, J., et al. (2015). 
Diagnostic accuracy of minimally invasive markers for detection of airway eosinophilia in asthma: 
a systematic review and meta-analysis. 
\emph{Lancet Respiratory Medicine.} \strong{3}: 290-300.
\url{https://doi.org/10.1016/S2213-2600(15)00050-8}
}
\keyword{datasets}
