\docType{package}
\name{dmetatools-package}
\alias{dmetatools}
\title{The 'dmetatools' package.}
\description{
Computational tools for meta-analysis of diagnostic accuracy test. This package enables computations of confidence interval for the AUC of summary ROC curve and some related AUC-based inference methods.
}
\author{Hisashi Noma <noma@ism.ac.jp>}
\references{
Noma, H. and Matsushima, Y. (2020). 
Confidence interval for the AUC of SROC curve and some related methods using bootstrap for meta-analysis of diagnostic accuracy studies. 
arXiv:2004.04339.
\url{https://arxiv.org/abs/2004.04339}
}
